+++ digitizedbooks/apps/publish/management/commands/boxrefresh.py
@@ -1,9 +1,9 @@
 from box import refresh_v2_token
-from publish.models import BoxToken
+from digitizedbooks.apps.publish.models import BoxToken
 from django.core.management.base import NoArgsCommand
 
 class Command(NoArgsCommand):
-    
+
     def handle_noargs(self, **options):
         token = BoxToken.objects.get(id=1)
         response = refresh_v2_token(token.client_id, token.client_secret, token.refresh_token)